docs(gallery): split plot_particle_systems into 4 particle examples
Gravitational swarm, fireworks, particle flow dynamics, and
magnetic-field iron filings now have individual gallery entries with
their own thumbnails, matching the rest of the split gallery.
